agents/requirements.txt
# Python Agent Services

langchain==0.1.20
langchain-openai==0.1.6
langchain-anthropic==0.1.11
langchain-community==0.0.38
openai==1.25.0
anthropic==0.25.0
redis==5.0.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.0
requests==2.31.0
duckduckgo-search==5.3.0
beautifulsoup4==4.12.3
lxml==5.2.1
